# Sync owned trials when calling `study.ask` and `study.get_trials`