--- a/tests/storages_tests/test_cached_storage.py
+++ b/tests/storages_tests/test_cached_storage.py
@@ -112,8 +112,21 @@ def test_read_trials_from_remote_storage() -> None:
         directions=[StudyDirection.MINIMIZE], study_name="test-study"
     )
 
-    storage.read_trials_from_remote_storage(study_id)
+    storage.read_trials_from_remote_storage(study_id, sync_owned_trials=False)
 
     # Non-existent study.
     with pytest.raises(KeyError):
-        storage.read_trials_from_remote_storage(study_id + 1)
+        storage.read_trials_from_remote_storage(study_id + 1, sync_owned_trials=False)
+
+    # Create a trial via CachedStorage and update it via backend storage directly.
+    trial_id = storage.create_new_trial(study_id)
+    base_storage.set_trial_param(
+        trial_id, "paramA", 1.2, optuna.distributions.FloatDistribution(-0.2, 2.3)
+    )
+    base_storage.set_trial_state_values(trial_id, TrialState.COMPLETE, values=[0.0])
+
+    storage.read_trials_from_remote_storage(study_id, sync_owned_trials=False)
+    assert storage.get_trial(trial_id).state == TrialState.RUNNING
+
+    storage.read_trials_from_remote_storage(study_id, sync_owned_trials=True)
+    assert storage.get_trial(trial_id).state == TrialState.COMPLETE