@c-bata 2023-03-24
FYI: I ran the following benchmark to compare how CachedStorage slows down with this change.

https://gist.github.com/c-bata/16d654471393f59e064482ffb8b52bf8

|  | master | This PR |
| --- | --- | --- |
| RDBStorage | 47.9438s | 51.0639s |
| CachedStorage(RDBStorage) | 36.6144s | 54.9168s |