@c-bata 2023-03-30
Let me unassign reviewers since the performance become slower than I expected. I will mark this ready for review after fixing the performance issue.